requirements.txt
pyserial
PyMySQL
datetime
pytz
apscheduler